全選都是假象

+++ course/api/course-doBatchIsDoDelete.php
@@ -3,7 +3,11 @@ require_once("db-connect.php");
 
 $id=$_POST["checkbox"];
 
-
+if(empty($id)){
+    echo "<script>alert('錯誤:未勾選項目')</script>";
+    echo "<script>history.go(-1)</script> ";
+    exit;
+}
 
 foreach($id as $value){
     $sql="DELETE FROM classes WHERE course_id='$value'";
